requirements.txt
streamlit
torch
torchvision
opencv-python
numpy
Pillow